+++ modules/meshcloud-replicator/replicator-management-account-access/outputs.tf
@@ -1,0 +1,4 @@
+output "management_account_role_arn" {
+  description = "Amazon Resource Name (ARN) of Management Account Role"
+  value = aws_iam_role.meshfed_service.arn
+}